+++ zetaclient/zetacore_observer.go
@@ -7,25 +7,21 @@ import (
 
 	"github.com/zeta-chain/zetacore/zetaclient/bitcoin"
 	"github.com/zeta-chain/zetacore/zetaclient/interfaces"
+	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 	"github.com/zeta-chain/zetacore/zetaclient/outtxprocessor"
 
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 
 	sdkmath "cosmossdk.io/math"
 
-	"github.com/pkg/errors"
-	prom "github.com/prometheus/client_golang/prometheus"
 	"github.com/rs/zerolog"
 	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	"github.com/zeta-chain/zetacore/zetaclient/config"
-	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 )
 
 const (
-	MaxLookaheadNonce   = 120
-	OutboundTxSignCount = "zetaclient_Outbound_tx_sign_count"
-	HotKeyBurnRate      = "zetaclient_hotkey_burn_rate"
+	MaxLookaheadNonce = 120
 )
 
 type ZetaCoreLog struct {
@@ -38,7 +34,6 @@ type CoreObserver struct {
 	bridge              interfaces.ZetaCoreBridger
 	signerMap           map[common.Chain]interfaces.ChainSigner
 	clientMap           map[common.Chain]interfaces.ChainClient
-	metrics             *metrics.Metrics
 	logger              ZetaCoreLog
 	cfg                 *config.Config
 	ts                  *metrics.TelemetryServer
@@ -51,7 +46,6 @@ func NewCoreObserver(
 	bridge interfaces.ZetaCoreBridger,
 	signerMap map[common.Chain]interfaces.ChainSigner,
 	clientMap map[common.Chain]interfaces.ChainClient,
-	metrics *metrics.Metrics,
 	logger zerolog.Logger,
 	cfg *config.Config,
 	ts *metrics.TelemetryServer,
@@ -73,16 +67,7 @@ func NewCoreObserver(
 	co.signerMap = signerMap
 
 	co.clientMap = clientMap
-	co.metrics = metrics
 	co.logger.ChainLogger.Info().Msg("starting core observer")
-	err := metrics.RegisterCounter(OutboundTxSignCount, "number of Outbound tx signed")
-	if err != nil {
-		co.logger.ChainLogger.Error().Err(err).Msg("error registering counter")
-	}
-	err = metrics.RegisterGauge(HotKeyBurnRate, "Fee burn rate of the hotkey")
-	if err != nil {
-		co.logger.ChainLogger.Error().Err(err).Msg("error registering gauge")
-	}
 	balance, err := bridge.GetZetaHotKeyBalance()
 	if err != nil {
 		co.logger.ChainLogger.Error().Err(err).Msg("error getting last balance of the hot key")
@@ -94,22 +79,6 @@ func NewCoreObserver(
 
 func (co *CoreObserver) Config() *config.Config {
 	return co.cfg
-}
-
-func (co *CoreObserver) GetPromCounter(name string) (prom.Counter, error) {
-	cnt, found := metrics.Counters[name]
-	if !found {
-		return nil, errors.New("counter not found")
-	}
-	return cnt, nil
-}
-
-func (co *CoreObserver) GetPromGauge(name string) (prom.Gauge, error) {
-	gauge, found := metrics.Gauges[name]
-	if !found {
-		return nil, errors.New("gauge not found")
-	}
-	return gauge, nil
 }
 
 func (co *CoreObserver) MonitorCore() {
@@ -170,12 +139,7 @@ func (co *CoreObserver) startCctxScheduler() {
 					}
 
 					// Set Current Hot key burn rate
-					gauge, err := co.GetPromGauge(HotKeyBurnRate)
-					if err != nil {
-						co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxEVM: failed to get prometheus gauge: %s for observer", metrics.PendingTxs)
-						continue
-					} // Gauge only takes float values
-					gauge.Set(float64(co.ts.HotKeyBurnRate.GetBurnRate().Int64()))
+					metrics.HotKeyBurnRate.Set(float64(co.ts.HotKeyBurnRate.GetBurnRate().Int64()))
 
 					// schedule keysign for pending cctxs on each chain
 					supportedChains := co.Config().GetEnabledChains()
@@ -196,12 +160,7 @@ func (co *CoreObserver) startCctxScheduler() {
 							continue
 						}
 						// Set Pending transactions prometheus gauge
-						gauge, err := ob.GetPromGauge(metrics.PendingTxs)
-						if err != nil {
-							co.logger.ZetaChainWatcher.Error().Err(err).Msgf("scheduleCctxEVM: failed to get prometheus gauge: %s for chain %d", metrics.PendingTxs, c.ChainId)
-							continue
-						}
-						gauge.Set(float64(totalPending))
+						metrics.PendingTxsPerChain.WithLabelValues(c.ChainName.String()).Set(float64(totalPending))
 
 						// #nosec G701 range is verified
 						zetaHeight := uint64(bn)