+++ changelog.md
@@ -6,9 +6,14 @@
 
 * [1789](https://github.com/zeta-chain/node/issues/1789) - block cross-chain transactions that involve restricted addresses
 
+### Refactor
+
+* [1783](https://github.com/zeta-chain/node/pull/1783) - refactor zetaclient metrics naming and structure
+
 ### Tests
 
 * [1767](https://github.com/zeta-chain/node/pull/1767) - add unit tests for emissions module begin blocker
+* [1791](https://github.com/zeta-chain/node/pull/1791) - add e2e tests for feature of restricted address
 
 ## Version: v13.0.0